refactor: Allow Surface implementers to override the intersect algo
