+++ crates/cherry-rs/src/core/mod.rs
@@ -1,6 +1,7 @@
 /// Data types for modeling ray tracing systems.
 pub(super) mod math;
 pub(crate) mod placement;
+pub(crate) mod ray;
 pub(crate) mod refractive_index;
 pub(crate) mod sequential_model;
 pub(crate) mod surfaces;
